pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discocry"
version = "1.0"
dependencies = [
"requests", 
"google-api-python-client",
"google-auth-httplib2",
"google-auth-oauthlib"
]

[tool.hatch.build.targets.wheel]
packages = ["emailclassifier"]
